Transaction Feature (#2)
* initiate select product

* create database for cart

* create checkout

* update transaction page

* add filter to transaction list

* update select product (animate clicked), add create expense, edit profile, and change password

* add trasition animation (animate navigation)

* add alert confirmation

+++ app/src/main/java/com/herlianzhang/mikropos/vo/Product.kt
@@ -7,7 +7,7 @@ import kotlinx.parcelize.Parcelize
 data class Product(
     val id: Int,
     val name: String?,
-    val price: Int?,
+    val price: Long?,
     val photo: String?,
     val totalStock: Int?
 ) : Parcelable
